snprintf error checking in auth_krb5.c
snprintf returns a negative value on error, but adding it to existing
plen won't make it necessarily go zero or negative, hence existing code
error checking was inadequate.

Use a temp variable to get the snprintf return add add it to plen in a
second step. Check the temp value if negative and plen for 0 or >=
pnamemax.